Refactor hydro pressure slice example by extracting utility functions
- Moved utility functions related to hydroelastic pressure slicing from `example_hydro_pressure_slice.py` to a new module `hydro_pressure_slice_utils.py`.
- Updated imports in `example_hydro_pressure_slice.py` to use the new utility functions.
- Removed redundant code from `example_hydro_pressure_slice.py` to improve readability and maintainability.
